LayoutTests/inspector/console/command-line-api-inspect-expected.txt

Tests that inspect() command line api works.


Running: testRevealElement


WebInspector.inspect called with: <p>
WebInspector.inspect's hints are: []
inspect($('p1')) = 
Selected node id: 'p1'.

Running: testRevealDatabase


WebInspector.inspect called with: Database
WebInspector.inspect's hints are: ["databaseId"]
inspect(db) = Database

Running: testRevealDatabase2


WebInspector.inspect called with: Database
WebInspector.inspect's hints are: ["databaseId"]
inspect(db2) = Database

Running: testRevealLocalStorage


WebInspector.inspect called with: Storage
WebInspector.inspect's hints are: ["domStorageId"]
inspect(localStorage) = Storage

Running: testRevealSessionStorage


WebInspector.inspect called with: Storage
WebInspector.inspect's hints are: ["domStorageId"]
inspect(sessionStorage) = Storage
